moved PersonalNoteViewSet class to view.py
