+++ notes/views.py
@@ -1,3 +1,16 @@
-from django.shortcuts import render
+from .models import PersonalNote
+from rest_framework import viewsets
+from .api import PersonalNoteSerializer
 
-# Create your views here.
+class PersonalNoteViewSet(viewsets.ModelViewSet):
+	serializer_class = PersonalNoteSerializer
+	queryset = PersonalNote.objects.none()
+
+	def get_queryset(self):
+		# import pdb; pdb.set_trace()
+		user = self.request.user
+
+		if user.is_anonymous:
+			return PersonalNote.objects.none()
+		else:
+			return PersonalNote.objects.filter(user=user)
